Tests/LibWeb/Text/expected/input-scrolls-caret-width-into-view.txt

scrolls to caret: true
rect=[18,8 1x16] color=rgb(0, 0, 0)
RTL scrolls to caret: true
